requirements.txt
pandas
openpyxl
catboost
scikit-learn
flask
numpy
statsmodels
scipy
psycopg2-binary
